@@ -3,15 +3,21 @@
     "description": "Extension for testing FastMap and ETL components with PHPUnit testing framework",
     "type": "library",
     "require": {
-        "php": "^8.1",
-        "phpunit/phpunit": "^9",
+        "php": "^8.2",
+        "phpunit/phpunit": "^10.0",
         "nikic/php-parser": "^4.10",
-        "php-etl/pipeline-contracts": "^0.3.0",
+        "php-etl/pipeline-contracts": "^0.4.0",
         "php-http/message": "^1.11",
         "php-http/mock-client": "^1.4@dev",
         "fakerphp/faker": "^1.19",
         "laminas/laminas-diactoros": "^2.6",
         "mikey179/vfsstream": "^1.6"
+    },
+    "require-dev": {
+        "phpstan/phpstan": "^1.10",
+        "friendsofphp/php-cs-fixer": "^3.0",
+        "infection/infection": "^0.26.18",
+        "rector/rector": "^0.15"
     },
     "license": "MIT",
     "authors": [
@@ -33,11 +39,15 @@
         }
     },
     "config": {
-        "bin-dir": "bin"
+        "bin-dir": "bin",
+        "allow-plugins": {
+            "php-http/discovery": true,
+            "infection/extension-installer": true
+        }
     },
     "extra": {
         "branch-alias": {
-            "dev-main": "0.4.x-dev"
+            "dev-main": "0.5.x-dev"
         }
     }
 }
