@@ -14,26 +14,26 @@ trait PipelineAssertTrait
 {
     abstract public static function assertThat($value, Constraint $constraint, string $message = ''): void;
 
-    protected function assertDoesIterateLike(iterable $expected, iterable $actual, $message = '')
+    protected function assertDoesIterateLike(iterable $expected, iterable $actual, $message = ''): void
     {
-        $this->assertThat($actual, new IteratesLike($expected, fn ($item) => new IsEqual($item)), $message);
+        static::assertThat($actual, new IteratesLike($expected, fn ($item) => new IsEqual($item)), $message);
     }
 
-    protected function assertDoesIterateExactly(iterable $expected, iterable $actual, $message = '')
+    protected function assertDoesIterateExactly(iterable $expected, iterable $actual, $message = ''): void
     {
-        $this->assertThat($actual, new IteratesLike($expected, fn ($item) => new IsIdentical($item)), $message);
+        static::assertThat($actual, new IteratesLike($expected, fn ($item) => new IsIdentical($item)), $message);
     }
 
-    protected function assertDoesNotIterateLike(iterable $expected, iterable $actual, $message = '')
+    protected function assertDoesNotIterateLike(iterable $expected, iterable $actual, $message = ''): void
     {
-        $this->assertThat($actual, new LogicalNot(
+        static::assertThat($actual, new LogicalNot(
             new IteratesLike($expected, fn ($item) => new IsEqual($item)),
         ), $message);
     }
 
-    protected function assertDoesNotIterateExactly(iterable $expected, iterable $actual, $message = '')
+    protected function assertDoesNotIterateExactly(iterable $expected, iterable $actual, $message = ''): void
     {
-        $this->assertThat($actual, new LogicalNot(
+        static::assertThat($actual, new LogicalNot(
             new IteratesLike($expected, fn ($item) => new IsIdentical($item)),
         ), $message);
     }
