@@ -22,21 +22,22 @@ final class PipelineLoadsLike extends Constraint
      * @param list<Type> $expected
      */
     public function __construct(
-        private iterable $source,
-        private iterable $expected,
+        private readonly iterable $source,
+        private readonly iterable $expected,
         callable $itemConstraintFactory,
-        private PipelineRunnerInterface $runner,
+        private readonly PipelineRunnerInterface $runner,
     ) {
         $this->itemConstraintFactory = $itemConstraintFactory;
     }
 
     /**
      * @param list<Type> $iterable
+     *
      * @return \Iterator<Type>
      */
     private function asIterator(iterable $iterable): \Iterator
     {
-        if (is_array($iterable)) {
+        if (\is_array($iterable)) {
             return new \ArrayIterator($iterable);
         }
         if (!$iterable instanceof \Iterator && $iterable instanceof \Traversable) {
@@ -98,7 +99,7 @@ final class PipelineLoadsLike extends Constraint
         foreach ($both as [$expectedItem, $actualItem]) {
             ++$index;
             $constraint = ($this->itemConstraintFactory)($expectedItem);
-            $constraint->evaluate($actualItem, sprintf("Values of Iteration #%d", $index)) !== true;
+            true !== $constraint->evaluate($actualItem, sprintf('Values of Iteration #%d', $index));
         }
 
         return true;
