@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Kiboko\Component\PHPUnitExtension\Mock;
 
-use Kiboko\Component\PHPUnitExtension\Mock;
 use Http\Mock\Client;
+use Kiboko\Component\PHPUnitExtension\Mock;
 use PhpParser\Builder;
 use PhpParser\Node;
 
@@ -15,7 +15,7 @@ final class HttpClientBuilder implements Builder
     private array $nodes;
 
     public function __construct(
-        private Mock\ResponseFactoryBuilder $responseFactory
+        private readonly Mock\ResponseFactoryBuilder $responseFactory
     ) {
     }
 
